Add concentrated liquidity pool details component and enhance price chart customization options.
